--- src/utils/importAll.js
@@ -1,7 +1,0 @@
-export function importAll(r) {
-  let images = {};
-  r.keys().map((item) => {
-    images[item.replace("./", "")] = r(item);
-  });
-  return images;
-}
